refactor(constructor): Refactor constructor entity public api on client
- Changed imports in constructor entities to public api on client
