+++ packages/client/__tests__/__mocks__/entities/constructor/Constructor.mock.ts
@@ -1,4 +1,4 @@
-import { Constructor } from '~entities/constructor/type';
+import { Constructor } from '~entities/constructor';
 
 export const ConstructorsMock: Constructor[] = [
     {
