+++ packages/client/src/entities/constructor/api/useConstructorServer.ts
@@ -1,11 +1,11 @@
 import { axios } from '~shared/api/axios';
 
-import { URLS } from '~entities/constructor/api/urls';
-import type { Constructor } from '~entities/constructor/type';
+import { CONSTRUCTOR_URLS } from '~entities/constructor/api';
+import type { Constructor } from '~entities/constructor';
 
 export async function useConstructorServer(ref: string) {
     try {
-        const { data: constructor } = await axios.get<Constructor>(URLS.ref(ref));
+        const { data: constructor } = await axios.get<Constructor>(CONSTRUCTOR_URLS.ref(ref));
 
         return { constructor };
     } catch (error: unknown) {
